docs(aepp): extract SpendCoins component
