+++ examples/browser/aepp/src/Basic.vue
@@ -23,54 +23,21 @@
     </div>
   </div>
 
-  <h2>Spend coins</h2>
-  <div class="group">
-    <div>
-      <div>Recipient address</div>
-      <div>
-        <input
-          v-model="spendTo"
-          placeholder="ak_..."
-        >
-      </div>
-    </div>
-    <div>
-      <div>Coins amount</div>
-      <div><input v-model="spendAmount"></div>
-    </div>
-    <div>
-      <div>Payload</div>
-      <div><input v-model="spendPayload"></div>
-    </div>
-    <button
-      :disabled="!aeSdk"
-      @click="spendPromise = spend()"
-    >
-      Spend
-    </button>
-    <div v-if="spendPromise">
-      <div>Spend result</div>
-      <Value :value="spendPromise" />
-    </div>
-  </div>
+  <SpendCoins />
 </template>
 
 <script>
 import { mapState } from 'vuex';
-import { encode, Encoding } from '@aeternity/aepp-sdk';
-import Value from './Value.vue';
+import Value from './components/Value.vue';
+import SpendCoins from './components/SpendCoins.vue';
 
 export default {
-  components: { Value },
+  components: { Value, SpendCoins },
   data: () => ({
     balancePromise: null,
     heightPromise: null,
     nodeInfoPromise: null,
     compilerVersionPromise: null,
-    spendTo: '',
-    spendAmount: '',
-    spendPayload: '',
-    spendPromise: null,
   }),
   computed: mapState(['aeSdk', 'address', 'networkId']),
   mounted() {
@@ -85,13 +52,6 @@ export default {
       },
       { immediate: true },
     );
-  },
-  methods: {
-    spend() {
-      return this.aeSdk.spend(this.spendAmount, this.spendTo, {
-        payload: encode(new TextEncoder().encode(this.spendPayload), Encoding.Bytearray),
-      });
-    },
   },
 };
 </script>
